Add CloudFormation template.
This is a significantly simplified version of what @al-amine created for the
DevOps IaC that manages the main deployment of this service. Much of the
simplification was what seemed to me to be obvious improvements, easier to
fix than to leave the original in, but some was simply removing what I
don't need for testing the ELK integration.