Get 'secrets' from Jenkins credential store, not parameters.
The problem is that any "triggered" build, such as a CI build triggered by a
Git push or a CD run triggered by CI completion, will *always* use the default
values for any parameters (https://issues.jenkins-ci.org/browse/JENKINS-30153).
The credential store seems not to have that limitation, so it's reasonable to
use it for details I've kept out of the scripts because they're sensitive or
semi-sensitive data.